+++ controller/handlers_test.go
@@ -1,3 +1,23 @@
+// Copyright (c) 2017 Northwestern Mutual.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
 package controller
 
 import (
@@ -144,6 +164,46 @@ func TestAddFunc(t *testing.T) {
 	assert.Nil(t, err)
 	assert.NotNil(t, result)
 
+}
+
+func TestUpdateFunc(t *testing.T) {
+	clearAllStores()
+	defer clearAllStores()
+	handlers := k8sEventHandler{}
+	setDecryptionKey(t)
+
+	spec.ProxyStore.Set(spec.APIProxy{
+		ObjectMeta: api.ObjectMeta{
+			Name:      "exampleAPIProxyOne",
+			Namespace: "foo",
+		},
+		Spec: spec.APIProxySpec{
+			Path: "api/v1/accounts",
+			Service: spec.Service{
+				Name: "my-service",
+				Port: 8080,
+			},
+		},
+	})
+	result, _ := spec.ProxyStore.Get("/api/v1/accounts")
+	assert.NotNil(t, result)
+	handlers.updateFunc(spec.APIProxy{
+		ObjectMeta: api.ObjectMeta{
+			Name:      "exampleAPIProxyOne",
+			Namespace: "foo",
+		},
+		Spec: spec.APIProxySpec{
+			Path: "/modified/path",
+			Service: spec.Service{
+				Name: "my-service",
+				Port: 8080,
+			},
+		},
+	})
+	result, _ = spec.ProxyStore.Get("/api/v1/accounts")
+	assert.Nil(t, result)
+	result, _ = spec.ProxyStore.Get("/modified/path")
+	assert.NotNil(t, result)
 }
 
 func clearAllStores() {