+++ spec/apikeybinding.go
@@ -158,6 +158,17 @@ func (n *subpathNode) doSetSubpath(pathSegments []string, subpath *Path) {
 	}
 }
 
+// Update will update an APIKeyBinding
+func (s *BindingFactory) Update(obj interface{}) error {
+	s.mutex.Lock()
+	defer s.mutex.Unlock()
+	binding, ok := obj.(APIKeyBinding)
+	if !ok {
+		return errors.New("grrr - you're only allowed add api key bindings to the api key binding store.... duh")
+	}
+	return s.set(binding)
+}
+
 // Set takes a APIKeyBinding and either adds it to the store
 // or updates it
 func (s *BindingFactory) Set(obj interface{}) error {
@@ -167,6 +178,10 @@ func (s *BindingFactory) Set(obj interface{}) error {
 	if !ok {
 		return errors.New("grrr - you're only allowed add api key bindings to the api key binding store.... duh")
 	}
+	return s.set(binding)
+}
+
+func (s *BindingFactory) set(binding APIKeyBinding) error {
 	logrus.Infof("Adding new APIKeyBinding named %s in namespace %s", binding.ObjectMeta.Name, binding.ObjectMeta.Namespace)
 	binding.hydrateSubpathTree()
 	if s.bindingMap[binding.ObjectMeta.Namespace] == nil {