+++ spec/store.go
@@ -24,6 +24,7 @@ package spec
 // a data structure
 type Store interface {
 	Set(obj interface{}) error
+	Update(obj interface{}) error
 	Get(params ...interface{}) (interface{}, error)
 	Delete(obj interface{}) (interface{}, error)
 	Clear()