+++ utils/utils_test.go
@@ -27,6 +27,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"k8s.io/kubernetes/pkg/api"
 )
 
 func TestComputeTargetPath(t *testing.T) {
@@ -120,4 +121,37 @@ func TestOmitHeaderValues(t *testing.T) {
 	copy = OmitHeaderValues(nil, "ommitted")
 	assert.Nil(copy, "map should be equal")
 
+}
+
+func TestCompareObjectMeta(t *testing.T) {
+	c1 := api.ObjectMeta{
+		Name:      "foo",
+		Namespace: "bar",
+	}
+	c2 := api.ObjectMeta{
+		Name:      "bar",
+		Namespace: "foo",
+	}
+	c3 := api.ObjectMeta{
+		Name:      "foo",
+		Namespace: "car",
+	}
+	c4 := api.ObjectMeta{
+		Name:      "bar",
+		Namespace: "car",
+	}
+
+	assert.True(t, CompareObjectMeta(c1, c1))
+	assert.False(t, CompareObjectMeta(c1, c2))
+	assert.False(t, CompareObjectMeta(c1, c3))
+	assert.False(t, CompareObjectMeta(c3, c4))
+}
+
+func TestNormalizePath(t *testing.T) {
+	assert.Equal(t, "/foo/bar", NormalizePath("foo////bar"))
+	assert.Equal(t, "/foo", NormalizePath("foo"))
+	assert.Equal(t, "/foo", NormalizePath("foo////"))
+	assert.Equal(t, "/foo/bar", NormalizePath("///foo////bar//"))
+	assert.Equal(t, "/", NormalizePath(""))
+	assert.Equal(t, "/", NormalizePath("////"))
 }